+++ RELEASENOTES.md
@@ -1,6 +1,7 @@
 # Release Notes
 
 ## Unreleased
+- Add support for Subject Name + Issuer (SNI) certificate authentication ([#46 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/46))
 - Update dependency package versions ([#47 by philon-msft](https://github.com/Azure/Microsoft.Azure.StackExchangeRedis/pull/47))
 
 ## 2.0.0